+++ NotificarePushKit/Sources/Internals/NotificarePushImpl.swift
@@ -64,12 +64,19 @@ internal class NotificarePushImpl: NSObject, NotificareModule, NotificarePush {
 
     internal func clearStorage() async throws {
         LocalStorage.clear()
+
+        _subscriptionStream.value = LocalStorage.subscription
+        _allowedUIStream.value = LocalStorage.allowedUI
     }
 
     internal func postLaunch() async throws {
         if hasRemoteNotificationsEnabled {
             logger.debug("Enabling remote notifications automatically.")
             try await updateDeviceSubscription()
+
+            if await hasNotificationPermission() {
+                await reloadActionCategories()
+            }
         }
     }