+++ Package.swift
@@ -56,63 +56,63 @@ let package = Package(
     targets: [
         .binaryTarget(
             name: "NotificareKit",
-            url: "https://cdn.notifica.re/libs/ios/4.1.0/spm-notificare.zip",
-            checksum: "7e79ad94aff0f7944b512b97aeb3d86ab18e2749b4465ce8973e09d133cc0b41"
+            url: "https://cdn.notifica.re/libs/ios/4.2.0/spm-notificare.zip",
+            checksum: "f5f0e8bd2ed80be9eecf6e6ed347be6c33e92fdaaea9893a7675a8dc3ce37dd0"
         ),
         .binaryTarget(
             name: "NotificareAssetsKit",
-            url: "https://cdn.notifica.re/libs/ios/4.1.0/spm-notificare-assets.zip",
-            checksum: "d4f8589b661b8a0ac6f23b462956be190552666ca4b84a2848c803c37151f970"
+            url: "https://cdn.notifica.re/libs/ios/4.2.0/spm-notificare-assets.zip",
+            checksum: "d7a205392ddfc5105b54bd14eee5f1f1e5a4ef1aab25ab155cb57c1ba23469bb"
         ),
         .binaryTarget(
             name: "NotificareGeoKit",
-            url: "https://cdn.notifica.re/libs/ios/4.1.0/spm-notificare-geo.zip",
-            checksum: "206941090b389a6aa51a538fafa22f1c0a9570c1eb69444bb2c0cd3735121312"
+            url: "https://cdn.notifica.re/libs/ios/4.2.0/spm-notificare-geo.zip",
+            checksum: "6b6baedd79cfc42f528a6c15693ee36a6ea92b65ecfb9e7cef20e57c27a34508"
         ),
         .binaryTarget(
             name: "NotificareInAppMessagingKit",
-            url: "https://cdn.notifica.re/libs/ios/4.1.0/spm-notificare-in-app-messaging.zip",
-            checksum: "84df16604f448d104d3b5175425dc76a7267e4f29fc0a45a27913090e4d29528"
+            url: "https://cdn.notifica.re/libs/ios/4.2.0/spm-notificare-in-app-messaging.zip",
+            checksum: "4d830a65f2323973c5bd416d8705501b71d33bb91df4d8b45663ffd25dbfcff1"
         ),
         .binaryTarget(
             name: "NotificareInboxKit",
-            url: "https://cdn.notifica.re/libs/ios/4.1.0/spm-notificare-inbox.zip",
-            checksum: "c02069a7bd95e313a600dde7af005e139d59edbad5a3939c0362d5520ce75d58"
+            url: "https://cdn.notifica.re/libs/ios/4.2.0/spm-notificare-inbox.zip",
+            checksum: "37b8f021bb6b9329031e7a553c86a3f249a05cfe33b72f60755638e0df7420d9"
         ),
         .binaryTarget(
             name: "NotificareLoyaltyKit",
-            url: "https://cdn.notifica.re/libs/ios/4.1.0/spm-notificare-loyalty.zip",
-            checksum: "1585c7927863c952fc459e55b23c039b50a9ede5c447e460aa2ad5784598b4fe"
+            url: "https://cdn.notifica.re/libs/ios/4.2.0/spm-notificare-loyalty.zip",
+            checksum: "db54b0ff2088f2f0ccf87b28bd002000db9a4e6f00aba4c8039eee547e6fd4d5"
         ),
         .binaryTarget(
             name: "NotificarePushKit",
-            url: "https://cdn.notifica.re/libs/ios/4.1.0/spm-notificare-push.zip",
-            checksum: "8385f410bb3b62d0fa34b4125a6fe86f8c0b4d829b6db3df2e875859b0870da9"
+            url: "https://cdn.notifica.re/libs/ios/4.2.0/spm-notificare-push.zip",
+            checksum: "47dd5bc07019e55bc2325ff6d00621eadf5c34578ddae706cf4f257b4f05210a"
         ),
         .binaryTarget(
             name: "NotificareNotificationServiceExtensionKit",
-            url: "https://cdn.notifica.re/libs/ios/4.1.0/spm-notificare-notification-service-extension.zip",
-            checksum: "c1f2d139da8c01a1c0fbcd48b57b965af333d214d8a7f0b686631f9a582ef10e"
+            url: "https://cdn.notifica.re/libs/ios/4.2.0/spm-notificare-notification-service-extension.zip",
+            checksum: "3eecb20a31b91da20e82bd24d756059da37c10b9d03f2e91b90c7af4961fdaf4"
         ),
         .binaryTarget(
             name: "NotificarePushUIKit",
-            url: "https://cdn.notifica.re/libs/ios/4.1.0/spm-notificare-push-ui.zip",
-            checksum: "2c29c665e646dac3efd97e42c692edaed1a6a82ceba62c129ab3cf8ca9c207d8"
+            url: "https://cdn.notifica.re/libs/ios/4.2.0/spm-notificare-push-ui.zip",
+            checksum: "36d37dac39c5b8e885976d5bcad902c13bc792db21580678f47aec2c1e9d7789"
         ),
         .binaryTarget(
             name: "NotificareScannablesKit",
-            url: "https://cdn.notifica.re/libs/ios/4.1.0/spm-notificare-scannables.zip",
-            checksum: "fef559268c930ee1e40ae8eaa51ef3f4b8c5e74abda2b3b368986d6b03728604"
+            url: "https://cdn.notifica.re/libs/ios/4.2.0/spm-notificare-scannables.zip",
+            checksum: "b503e047c909f4827ce28ae674ee6c9f0c9484477065475379bf2031e87add04"
         ),
         .binaryTarget(
             name: "NotificareUserInboxKit",
-            url: "https://cdn.notifica.re/libs/ios/4.1.0/spm-notificare-user-inbox.zip",
-            checksum: "1875e420f06f890996c3088d11e2483bb1e362918a7d90492f7695b93287a5f0"
+            url: "https://cdn.notifica.re/libs/ios/4.2.0/spm-notificare-user-inbox.zip",
+            checksum: "f8960e1acf91fedbf63da2a55003c75ac1115bcb18984047f84f5a3296b90d9e"
         ),
         .binaryTarget(
             name: "NotificareUtilitiesKit",
-            url: "https://cdn.notifica.re/libs/ios/4.1.0/spm-notificare-utilities.zip",
-            checksum: "52ccf49487514c7775c700f2cd23581d66bc6bd6ec834e358588a5dcf91d580f"
+            url: "https://cdn.notifica.re/libs/ios/4.2.0/spm-notificare-utilities.zip",
+            checksum: "6fe23dbb65ce020bbee6d930ec4fe4e825fbe1f6ed07033b027bd3d55162f80f"
         ),
     ]
 )